requirements.txt
# Dependencias del proyecto DB-MANAGER

# Base de datos PostgreSQL
psycopg2-binary>=2.9.0

# Generación de datos falsos (opcional pero recomendado)
faker>=18.0.0

# GUI (Tkinter viene con Python, pero por si acaso)
# tkinter

# Testing (opcional)
# pytest>=7.0.0
